@@ -4,7 +4,8 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at https://mozilla.org/MPL/2.0/.
 
-//+build go1.7
+//go:build go1.7
+// +build go1.7
 
 package toplevel
 
