@@ -4,7 +4,8 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at http://mozilla.org/MPL/2.0/.
 
-// +build go1.14,!noquic,!go1.17
+//go:build go1.15 && !noquic
+// +build go1.15,!noquic
 
 package connections
 
@@ -13,10 +14,8 @@ import (
 	"crypto/tls"
 	"net"
 	"net/url"
-	"strings"
 	"sync"
 	"sync/atomic"
-	"syscall"
 	"time"
 
 	"github.com/lucas-clemente/quic-go"
@@ -81,7 +80,7 @@ func (t *quicListener) OnExternalAddressChanged(address *stun.Host, via string) 
 }
 
 func (t *quicListener) serve(ctx context.Context) error {
-	network := strings.ReplaceAll(t.uri.Scheme, "quic", "udp")
+	network := quicNetwork(t.uri)
 
 	udpAddr, err := net.ResolveUDPAddr(network, t.uri.Host)
 	if err != nil {
@@ -99,25 +98,12 @@ func (t *quicListener) serve(ctx context.Context) error {
 	svc, conn := stun.New(t.cfg, t, udpConn)
 	defer conn.Close()
 
-	quicWrapper := quicWrapper{
-		PacketConn: conn,
-		underlying: udpConn,
-	}
-	var wrapped net.PacketConn = &quicWrapper
-
-	if oobConn, ok := conn.(oobConn); ok {
-		l.Debugf("wrapping in oob conn")
-		wrapped = &oobConnWrapper{
-			quicWrapper, oobConn,
-		}
-	}
-
 	go svc.Serve(ctx)
 
-	registry.Register(t.uri.Scheme, wrapped)
-	defer registry.Unregister(t.uri.Scheme, wrapped)
+	registry.Register(t.uri.Scheme, conn)
+	defer registry.Unregister(t.uri.Scheme, conn)
 
-	listener, err := quic.Listen(wrapped, t.tlsCfg, quicConfig)
+	listener, err := quic.Listen(conn, t.tlsCfg, quicConfig)
 	if err != nil {
 		l.Infoln("Listen (BEP/quic):", err)
 		return err
@@ -204,7 +190,7 @@ func (t *quicListener) LANAddresses() []*url.URL {
 	uri := maybeReplacePort(t.uri, t.laddr)
 	t.mut.Unlock()
 	addrs := []*url.URL{uri}
-	network := strings.ReplaceAll(uri.Scheme, "quic", "udp")
+	network := quicNetwork(uri)
 	addrs = append(addrs, getURLsForAllAdaptersIfUnspecified(network, uri)...)
 	return addrs
 }
@@ -246,34 +232,4 @@ func (f *quicListenerFactory) New(uri *url.URL, cfg config.Wrapper, tlsCfg *tls.
 
 func (quicListenerFactory) Enabled(cfg config.Configuration) bool {
 	return true
-}
-
-// quicWrapper provides methods used by quic
-// https://github.com/lucas-clemente/quic-go/blob/master/packet_handler_map.go#L85
-type quicWrapper struct {
-	net.PacketConn
-	underlying *net.UDPConn
-}
-
-// SetReadBuffer is required by QUIC
-func (s *quicWrapper) SetReadBuffer(size int) error {
-	return s.underlying.SetReadBuffer(size)
-}
-
-// SyscallConn is required by QUIC
-func (s *quicWrapper) SyscallConn() (syscall.RawConn, error) {
-	return s.underlying.SyscallConn()
-}
-
-// oobConn is used to assert that stun package returned a net.PacketConn that implements this interface.
-// If it does, we then wrap quicWrapper in oobConnWrapper, to expose those methods to QUIC package.
-type oobConn interface {
-	ReadMsgUDP(b, oob []byte) (n, oobn, flags int, addr *net.UDPAddr, err error)
-	WriteMsgUDP(b, oob []byte, addr *net.UDPAddr) (n, oobn int, err error)
-}
-
-// See: https://pkg.go.dev/github.com/lucas-clemente/quic-go#OOBCapablePacketConn
-type oobConnWrapper struct {
-	quicWrapper
-	oobConn
 }
