@@ -4,13 +4,15 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at http://mozilla.org/MPL/2.0/.
 
-// +build go1.14,!noquic,!go1.17
+//go:build go1.15 && !noquic
+// +build go1.15,!noquic
 
 package connections
 
 import (
 	"crypto/tls"
 	"net"
+	"net/url"
 
 	"github.com/lucas-clemente/quic-go"
 	"github.com/syncthing/syncthing/lib/util"
@@ -22,6 +24,17 @@ var (
 		KeepAlive:          true,
 	}
 )
+
+func quicNetwork(uri *url.URL) string {
+	switch uri.Scheme {
+	case "quic4":
+		return "udp4"
+	case "quic6":
+		return "udp6"
+	default:
+		return "udp"
+	}
+}
 
 type quicTlsConn struct {
 	quic.Session
@@ -47,21 +60,7 @@ func (q *quicTlsConn) Close() error {
 }
 
 func (q *quicTlsConn) ConnectionState() tls.ConnectionState {
-	qcs := q.Session.ConnectionState()
-	return tls.ConnectionState{
-		Version:                     qcs.Version,
-		HandshakeComplete:           qcs.HandshakeComplete,
-		DidResume:                   qcs.DidResume,
-		CipherSuite:                 qcs.CipherSuite,
-		NegotiatedProtocol:          qcs.NegotiatedProtocol,
-		NegotiatedProtocolIsMutual:  qcs.NegotiatedProtocolIsMutual,
-		ServerName:                  qcs.ServerName,
-		PeerCertificates:            qcs.PeerCertificates,
-		VerifiedChains:              qcs.VerifiedChains,
-		SignedCertificateTimestamps: qcs.SignedCertificateTimestamps,
-		OCSPResponse:                qcs.OCSPResponse,
-		TLSUnique:                   qcs.TLSUnique,
-	}
+	return q.Session.ConnectionState().TLS.ConnectionState
 }
 
 // Sort available packet connections by ip address, preferring unspecified local address.
