@@ -834,6 +834,7 @@ angular.module('syncthing.core')
                 }
                 $scope.currentSharing.selected[n.deviceID] = true;
             });
+            $scope.currentSharing.shared.sort(deviceCompare);
             $scope.currentSharing.unrelated = $scope.deviceList().filter(function (n) {
                 return n.deviceID !== $scope.myID && !$scope.currentSharing.selected[n.deviceID];
             });
@@ -2670,15 +2671,18 @@ angular.module('syncthing.core')
             };
             switch (type) {
                 case "override":
-                    params.heading = $translate.instant("Override");
+                    params.heading = $translate.instant("Override Changes");
+                    params.icon = "fas fa-arrow-circle-up"
                     params.operation = "override";
                     break;
                 case "revert":
-                    params.heading = $translate.instant("Revert");
+                    params.heading = $translate.instant("Revert Local Changes");
+                    params.icon = "fas fa-arrow-circle-down"
                     params.operation = "revert";
                     break;
                 case "deleteEnc":
                     params.heading = $translate.instant("Delete Unexpected Items");
+                    params.icon = "fas fa-minus-circle"
                     params.operation = "revert";
                     break;
             }
